Move resolveRepositories function to new file
